Added AOC 2023 Day 19
I can feel the difficulty rising. Today's input was kind of difficult to parse. Anyways part1 was easy. Part2 had to go and use interval magic which we used in Day 5.  Had to rewrite and clean up code written in temp.cpp. Implementing a proper go solution is becoming tougher for me😢.
This is just a bookmark so that I can come and look at it later.