Added AOC 2023 Day 20
Today's input parsing and arranging made me a take a lot of time. Today's solution part1 was straightforward solution. But part 2 required some hardcoding as I had to figure out that the parent of rx was a conjunction module which had other conjunction module as inputs and we have to do lcm to get cycle lengths for repetition of a low pulse etc. It was all because of Input which made my solution work like Day 8. I have to come back and look back and see if I can get better solution with any insights later on.
This is just a bookmark so that I can come and look at it later.