Added AOC 2023 Day 21
Today's part 1 was easy, A simple BFS. Part2 solution was kind of like magic which was made possible because of input. I wanted to implement a solution which makes sense. The part2 was just identifying that the number of garden plots is quadratic with f(n) = 131*n + 65 which we could understand using math but wasn't intuitive to me. Have to revisit one day and see if I can use another solution for this day.
This is just a bookmark so that I can come and look at it later.