Added AOC 2023 Day 22
Today's solution was really a big brain one. I tried brute force initially which was intuitive for me but takes a lot of while to run for part 2. So sorted and simulated based on z coordinate and part2 was just simulation using topological sort which was interesting af. I learnt so much today. 3D Simulation was kind of hard to visualize and debug for part1 but learnt a lot.  I feel this and yesterday's were the most difficult days for me this year.
This is just a bookmark so that I can come and look at it later.