Added AOC 2023 Day 23
Today's was kind of easier after those 2 hellish days. Part 1 was easily manageable with  bruteforce. Idk but part2 for my C++ solution was taking > 5 minutes. So I had to use another method to find the longest path. I tried to condense the grid into a graph kind of structure which I learnt was edge contraction/graph condensation which allowed me to get the longest path fast, this was also possible because we don't want to turn back which made it all possible. Today's was interesting as I see some people manage to solve it using bruteforce using Rust etc which was surprising .
This is just a bookmark so that I can come and look at it later.