Added AOC 2023 Day 24
Day 24 was one of the harder days. Part1 was doable with normal math. But Part 2 was hard. It required a lot of math and code debugging. I initially solved the solution using z3,SymPy, mathematica etc as everyone did because it was kind of hard to do it. I then took some days off and revisited it and wanted to solve without using some blackbox magic solvers. I then just did the normally linear algebra stuff to get the answer but I'm having a precision problem which might because of using double and calculating inverse using determinant formula. I will come back to it later and use some of the numerical stable ones like LU decomposition. Or might use some other method using eigen vectors, plane equations, cross product, basis magic
This is just a bookmark so that I can come and look at it later.